recipes/docker/helper/requirements.txt
confluent-kafka
Faker
numpy
pandas
typer
pyyaml
requests
